+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Piano SDK for Android
 
-## v2.12.0-SNAPSHOT
+## v2.12.0
 * Add device restriction support
 * BREAKING CHANGE: PianoId initialization requires Context now
 * BREAKING CHANGE: ShowTemplateController requires DeviceIdProvider now